Allow connections over both IPv4 and IPv6 to the same socket
As long as the server is instantiated with new WebSocketServer(IpAddress.IPv6Any, port), this change will allow the server to accept connections over both IPv4 and IPv6. Note that there are better ways to do it under the Framework 4.5, but given that websocket-sharp is compiled under 3.5, we need this little trick to accomplish the dual-stack behaviour.