The program will be blocked when we use the properties of isAlive ActiveID and InactiveID and all methods relatived to ping undering the case of any remote client was closed suddenly.
That's a big problem when we use the software in real time data transferring and the network was interrupted unexpectedly, so I
change them to asynchronous method with setting macro definition of
__PING_ASYNC at property of project

1.Steps to reproduce:
Start WebSocketServer
Start two or more clients to connect to WebSocket Server
Unplugged the net cable of one of clients' manually
The property of isAlive ActiveID or InActiveID was used before sending message in WebSocketBehavior or the event of _sweepTimer was triggered at occasion of sending message.

2.Actual result:
All other rest clients could not receiving any data from WebSocket Server in 10 to 30 seconds until the WebSocket Server receiving the event of that client was closed or offline.

3.Expected result:
Send messages quickly and continusly

4.Proposed solution:
Replacing the usage of the ping method with pingAsync added by me